requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0

# OpenAI and embeddings
openai>=1.0.0
tiktoken>=0.5.0

# Vector storage and retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Evaluation
ragas>=0.1.0
datasets>=2.14.0
nltk>=3.8.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# For development
pytest>=7.4.0
black>=23.0.0